[ucd/normal] Drop tables module
